+++ Addon/ElvUI_mMediaTag/options/changelog/4.01.lua
@@ -1,0 +1,19 @@
+local mMT, DB, M, E, P, L, MEDIA = unpack(ElvUI_mMediaTag)
+
+mMT.Changelog[401] = {
+	DATE = "17.03.2026",
+	IMPORTANT = {
+		"[INFO]: A completely new update for Midnight, almost everything has been optimized and restructured.",
+		"[INFO]: Some modules and features are still missing, but this is a fully functional and working version for Midnight. There will be no version for Classic, where the 3.xx version should still work. I will continue to add more features over time.",
+	},
+	UPDATE = {
+		"[System]: Update Logo & Icon.",
+		"[Interrupt-On-CD]: Rework and Optimized, it should now work correctly.",
+	},
+    NEW = {
+        "[TAGs]: New Health Tags hide out of combat and no decimal.",
+        "[Class-Icons]: Add mMT Class Icons to Details.",
+        "[Details-Embedded]: Details Embedded module is back.",
+        "[Auto-Quest]: Auto accept and tur in Quests."
+    }
+}